requirements.txt
openai
python-dotenv
flask
flask-cors
gunicorn